requirements.txt
anytree==2.12.1
click==8.1.8
Flask==3.1.0
matplotlib==3.10.1
numpy==2.2.4
openai==1.68.2
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
torch==2.6.0